other_projects/Open_Brush_AI/pyproject.toml
[project]
name = "open-brush-ai"
version = "0.1.0"
description = "AI-powered painting controller for Open Brush VR"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.80.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[project.scripts]
open-brush-ai = "open_brush_ai.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
